refactor(config-ui): adjust the review comment
