+++ config-ui/src/pages/blueprint/create/step-one/example/gitlab.ts
