+++ config-ui/src/pages/blueprint/create/step-one/example/jenkins.ts
