+++ config-ui/src/utils/operator.ts
@@ -25,6 +25,14 @@ export type OperateConfig = {
   formatReason?: (err: unknown) => string
 }
 
+/**
+ * 
+ * @param request -> a request
+ * @param config 
+ * @param config.setOperating -> Control the status of the request
+ * @parma config.formatReason -> Show the reason for the failure
+ * @returns 
+ */
 export const operator = async <T>(
   request: () => Promise<T>,
   config?: OperateConfig
